Move sec council messages
